feat(Contract): contract settings
